+++ Runtime/Shaders/WaterCommon.hlsl
@@ -11,35 +11,35 @@
 #define WATER_TIME _Time.y
 
 #define DEPTH_MULTIPLIER 1 / _MaxDepth
-#define WaterBufferA(uv) SAMPLE_TEXTURE2D(_WaterBufferA, sampler_ScreenTextures_linear_clamp, half2(uv.x, 1-uv.y))
-#define WaterBufferAVert(uv) SAMPLE_TEXTURE2D_LOD(_WaterBufferA, sampler_ScreenTextures_linear_clamp, half2(uv.x, 1-uv.y), 0)
+#define WaterBufferA(uv) SAMPLE_TEXTURE2D(_WaterBufferA, sampler_ScreenTextures_linear_clamp, float2(uv.x, 1-uv.y))
+#define WaterBufferAVert(uv) SAMPLE_TEXTURE2D_LOD(_WaterBufferA, sampler_ScreenTextures_linear_clamp, float2(uv.x, 1-uv.y), 0)
 
 ///////////////////////////////////////////////////////////////////////////////
 //          	   	      Water shading functions                            //
 ///////////////////////////////////////////////////////////////////////////////
 
-half3 Scattering(half depth)
+float3 Scattering(float depth)
 {
-    const half grad = saturate(exp2(-depth * DEPTH_MULTIPLIER));
+    const float grad = saturate(exp2(-depth * DEPTH_MULTIPLIER));
     return _ScatteringColor * (1 - grad);
 }
 
-float4 DetailUVs(float3 positionWS, half noise)
+float4 DetailUVs(float3 positionWS, float noise)
 {
-    float4 output = positionWS.xzxz * half4(0.4, 0.4, 0.1, 0.1);
+    float4 output = positionWS.xzxz * float4(0.4, 0.4, 0.1, 0.1);
     output.xy -= WATER_TIME * 0.1h * 2 + (noise * 0.2); // small detail
     output.zw += WATER_TIME * 0.05h * 2 + (noise * 0.1); // medium detail
     return output;
 }
 
-void DetailNormals(inout float3 normalWS, float4 uvs, half4 waterFX)
+void DetailNormals(inout float3 normalWS, float4 uvs, float4 waterFX)
 {
-    half2 detailBump1 = SAMPLE_TEXTURE2D(_SurfaceMap, sampler_SurfaceMap, uvs.zw * 0.2).xy * 2 - 1;
-    half2 detailBump2 = SAMPLE_TEXTURE2D(_SurfaceMap, sampler_SurfaceMap, uvs.xy*0.2).xy * 2 - 1;
-    half2 detailBump = (detailBump1 + detailBump2 * 0.5) * saturate(1000 * 0.25);
+    float2 detailBump1 = SAMPLE_TEXTURE2D(_SurfaceMap, sampler_SurfaceMap, uvs.zw * 0.2).xy * 2 - 1;
+    float2 detailBump2 = SAMPLE_TEXTURE2D(_SurfaceMap, sampler_SurfaceMap, uvs.xy*0.2).xy * 2 - 1;
+    float2 detailBump = (detailBump1 + detailBump2 * 0.5) * saturate(1000 * 0.25);
 
-    half3 normal1 = half3(detailBump.x, 0, detailBump.y) * _BoatAttack_Water_MicroWaveIntensity;
-    half3 normal2 = half3(1 - waterFX.y, 0.5h, 1 - waterFX.z) - 0.5;
+    float3 normal1 = float3(detailBump.x, 0, detailBump.y) * _BoatAttack_Water_MicroWaveIntensity;
+    float3 normal2 = float3(1 - waterFX.y, 0.5h, 1 - waterFX.z) - 0.5;
     normalWS = normalize(normalWS + normal1 + normal2);
 }
 
@@ -50,7 +50,7 @@ Varyings WaveVertexOperations(Varyings input)
 
     input.uv = DetailUVs(input.positionWS, input.fogFactorNoise.y);
 
-    half4 screenUV = ComputeScreenPos(TransformWorldToHClip(input.positionWS));
+    float4 screenUV = ComputeScreenPos(TransformWorldToHClip(input.positionWS));
     screenUV.xyz /= screenUV.w;
 
     WaveStruct wave;
@@ -67,8 +67,8 @@ Varyings WaveVertexOperations(Varyings input)
     input.preWaveSP = screenUV.xyz; // pre-displaced screenUVs
 
     // distance blend
-    half distanceBlend = saturate(abs(length((_WorldSpaceCameraPos.xz - input.positionWS.xz) * 0.005)) - 0.25);
-    input.normalWS = lerp(input.normalWS, half3(0, 1, 0), distanceBlend);
+    float distanceBlend = saturate(abs(length((_WorldSpaceCameraPos.xz - input.positionWS.xz) * 0.005)) - 0.25);
+    input.normalWS = lerp(input.normalWS, float3(0, 1, 0), distanceBlend);
 
     return input;
 }
@@ -92,21 +92,21 @@ void InitializeInputData(Varyings input, out WaterInputData inputData, float2 sc
 
 float3 WaterShading(WaterInputData input, float2 screenUV)
 {
-    half fresnelTerm = CalculateFresnelTerm(input.normalWS, input.viewDirectionWS);
+    float fresnelTerm = CalculateFresnelTerm(input.normalWS, input.viewDirectionWS);
     Light mainLight = GetMainLight(TransformWorldToShadowCoord(input.positionWS), input.positionWS, 0);
 
-    half3 GI = SampleSH(input.normalWS) * 1.5;
+    float3 GI = SampleSH(input.normalWS) * 1.5;
 
     BRDFData brdfData;
-    half alpha = 1;
-    InitializeBRDFData(half3(0, 0, 0), 0, half3(1, 1, 1), 1, alpha, brdfData);
-    half3 spec = DirectBDRF(brdfData, input.normalWS, mainLight.direction, input.viewDirectionWS) * mainLight.color * 10;
+    float alpha = 1;
+    InitializeBRDFData(float3(0, 0, 0), 0, float3(1, 1, 1), 1, alpha, brdfData);
+    float3 spec = DirectBDRF(brdfData, input.normalWS, mainLight.direction, input.viewDirectionWS) * mainLight.color * 10;
 
-    half3 sss = GI;
+    float3 sss = GI;
     sss *= Scattering(1000);
 
-    half3 reflection = SampleReflections(input.normalWS, input.viewDirectionWS, screenUV, 0.0);
-    half3 output = lerp(lerp(sss, reflection + spec, fresnelTerm), 0, 0);
+    float3 reflection = SampleReflections(input.normalWS, input.viewDirectionWS, screenUV, 0.0);
+    float3 output = lerp(lerp(sss, reflection + spec, fresnelTerm), 0, 0);
     output = MixFog(output, input.fogCoord);
     return output;
 }
@@ -134,7 +134,7 @@ Varyings WaterVertex(Attributes v)
 }
 
 // Fragment for water
-half4 WaterFragment(Varyings IN) : SV_Target
+float4 WaterFragment(Varyings IN) : SV_Target
 {
     float4 screenUV = 0.0;
     screenUV.zw = IN.preWaveSP.xy;
@@ -142,7 +142,7 @@ half4 WaterFragment(Varyings IN) : SV_Target
     WaterInputData inputData;
     InitializeInputData(IN, inputData, screenUV.xy);
 
-    half4 current;
+    float4 current;
     current.a = WaterNearFade(IN.positionWS);
     current.rgb = WaterShading(inputData, screenUV.xy);
